added policy for individual controller and edited rspec

+++ app/controllers/todos_controller.rb
@@ -1,19 +1,22 @@
 class TodosController < ApplicationController
   before_action :fetch_todo_list
-  before_action :authorize_todo_list
+  # before_action :authorize_todo_list
   before_action :set_todo, only: %i[show update destroy]
 
   def index
+    authorize @todo_list, :view_todos?
     todos = @todo_list.todos
     render json: todos, status: :ok
   end
 
   def show
+    authorize @todo
     render json: @todo, status: :ok
   end
 
   def create
     todo = @todo_list.todos.build(todo_params)
+    authorize todo
     if todo.save
       render json: { todo:, message: 'Todo created successfully' }, status: :created
     else
@@ -22,6 +25,7 @@ class TodosController < ApplicationController
   end
 
   def update
+    authorize @todo
     if @todo.update(todo_params)
       render json: { todo: @todo, message: 'Todo updated successfully' }, status: :ok
     else
@@ -30,6 +34,7 @@ class TodosController < ApplicationController
   end
 
   def destroy
+    authorize @todo
     if @todo.destroy
       render json: { message: 'Todo deleted successfully' }, status: :ok
     else
@@ -46,10 +51,6 @@ class TodosController < ApplicationController
   def fetch_todo_list
     @todo_list = TodoList.find_by(id: params[:todo_list_id])
     render json: { error: 'No TodoList exists' }, status: :not_found unless @todo_list
-  end
-
-  def authorize_todo_list
-    authorize @todo_list, :access_todos?
   end
 
   def set_todo
